# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /data:
    get:
      tags: [Data]
      operationId: getUserData
      summary: Get user's planner data
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: User data returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPlannerData'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    put:
      tags: [Data]
      operationId: putUserData
      summary: Save user's planner data
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPlannerData'
      responses:
        '200':
          description: Data saved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPlannerData'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — `Bearer <sid>`.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    UserPlannerData:
      type: object
      properties:
        currentWorkouts:
          type: array
          items: {}
        historyEntries:
          type: array
          items: {}
        currentReflection:
          nullable: true
        reflectionShownCount:
          type: integer
        motivationalDate:
          type: [string, 'null']
        motivationalLabel:
          type: [string, 'null']