# lib/api-zod/src/generated/types/userPlannerData.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserPlannerData {
  currentWorkouts?: unknown[];
  historyEntries?: unknown[];
  currentReflection?: unknown | null;
  reflectionShownCount?: number;
  /** @nullable */
  motivationalDate?: string | null;
  /** @nullable */
  motivationalLabel?: string | null;
}